Do not return an error when the profile is empty
