+++ relay-profiling/src/error.rs
@@ -1,0 +1,17 @@
+use failure::Fail;
+
+#[derive(Debug, Fail)]
+pub enum ProfileError {
+    #[fail(display = "invalid json in profile")]
+    InvalidJson(#[cause] serde_json::Error),
+    #[fail(display = "invalid base64 value")]
+    InvalidBase64Value,
+    #[fail(display = "invalid sampled profile")]
+    InvalidSampledProfile,
+    #[fail(display = "cannot serialize payload")]
+    CannotSerializePayload,
+    #[fail(display = "not enough samples")]
+    NotEnoughSamples,
+    #[fail(display = "platform not supported")]
+    PlatformNotSupported,
+}
